program read write binary update
*program metadata which includes c function symbols are added to .bin
generated file
*struct member types are rearranged

+++ include/tasm/tasm_ast.h
@@ -115,8 +115,8 @@ typedef struct tasm_ast{
 
         // tile c interface
         struct ast_cfunction {
-            int ret_type;
-            int* arg_types;
+            uint8_t ret_type;
+            uint8_t* arg_types;
             size_t arg_count;
             const char* name;
         } cfunction;
